pyproject.toml
[tool.poetry]
name = "riley-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
bs4 = "^0.0.1"
discord = "^1.0.1"
flask = "^1.1.2"
praw = "^7.1.0"
lavalink = "^3.1.2"
wavelink = "^0.9.4"
requests-html = "^0.10.0"
wikipedia = "^1.4.0"
search-engine-parser = "^0.6.1"
better-profanity = "^0.6.1"
gsearch = "^1.6.0"
youtube-dl = "^2020.9.20"
python-resize-image = "^1.1.19"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
